Mod from patch in 895550cdec026559c8a519b2010b42a487dee5fa somehow got left out

+++ cellprofiler/modules/measureimagequality.py
@@ -788,7 +788,7 @@ class MeasureImageQuality(cpm.CPModule):
                 grid = i * n + j + 1
                 if image.has_mask:
                     grid[np.logical_not(image.mask)] = 0
-                grid_range = np.arange(0, m*n+1)
+                grid_range = np.arange(0, m*n+1,dtype=np.int32)
                 #
                 # Do the math per label
                 #
